change int to long in date constructor

+++ src/test/java/dbus/TimelineTest.java
@@ -12,11 +12,11 @@ import static org.assertj.core.api.Assertions.assertThat;
 public abstract class TimelineTest<E extends Date, T extends Timeline<E>> {
 
     abstract T construct(List<E> input);
-    abstract E date(int time);
+    abstract E date(long time);
 
     private final T timeline = construct(dates(0, 5, 9));
 
-    private List<E> dates(int... times) {
+    private List<E> dates(long... times) {
         return Arrays.stream(times)
                 .mapToObj(this::date)
                 .collect(Collectors.toList());
